requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
Pillow
torch
torchvision
timm
albumentations
lightning